requirements.txt
flask
Flask-SQLAlchemy
mysql-connector-python
pydantic
python-dotenv
web3
pytest
flasgger
Flask-Limiter
gunicorn
